Moving core-metadata to cmd
