+++ Makefile
@@ -5,18 +5,22 @@ EXPORT_DISTRO_VERSION=$(shell cat cmd/export-distro/VERSION)
 CORE_DATA_VERSION=$(shell cat cmd/core-data/VERSION)
 CORE_COMMAND_VERSION=$(shell cat cmd/core-command/VERSION)
 
-MICROSERVICES=cmd/core-data/core-data cmd/core-command/core-command cmd/export-client/export-client cmd/export-distro/export-distro
+MICROSERVICES=cmd/core-data/core-data cmd/core-metadata/core-metadata \
+	cmd/core-command/core-command cmd/export-client/export-client \
+	cmd/export-distro/export-distro
 .PHONY: $(MICROSERVICES)
 
 
 build: $(MICROSERVICES)
-	echo go build ./core/metadata
 
 cmd/core-data/core-data:
-	go build -ldflags "-X main.version=$(CORE_DATA_VERSION)" -o cmd/core-data/core-data ./cmd/core-data 
+	go build -ldflags "-X main.version=$(CORE_DATA_VERSION)" -o cmd/core-data/core-data ./cmd/core-data
+
+cmd/core-metadata/core-metadata:
+	go build -ldflags "-X main.version=$(CORE_METADATA_VERSION)" -o cmd/core-metadata/core-metadata ./cmd/core-metadata
 
 cmd/core-command/core-command:
-	go build -ldflags "-X main.version=$(CORE_COMMAND_VERSION)" -o cmd/core-command/core-command ./cmd/core-command 
+	go build -ldflags "-X main.version=$(CORE_COMMAND_VERSION)" -o cmd/core-command/core-command ./cmd/core-command
 
 cmd/export-client/export-client:
 	go build -ldflags "-X main.version=$(EXPORT_CLIENT_VERSION)" -o cmd/export-client/export-client ./cmd/export-client
