+++ cmd/core-metadata/res/configuration-docker.json
