+++ cmd/core-metadata/res/configuration.json
